feat: Add utility function `cn` for merging CSS classes
This commit adds a new utility function `cn` to the `utils.ts` file. The `cn` function takes in multiple CSS class values and merges them using the `clsx` and `twMerge` libraries. This utility function will be useful for dynamically combining CSS classes in components.

It also includes the addition of the `clsx` and `twMerge` dependencies in the `package.json` file.

Note: The recent commits are not relevant to the current changes.